+++ src/meshblock.hpp
@@ -3,6 +3,7 @@
 
 #include "vec3.hpp"
 #include "ray.hpp"
+#include <stdexcept>
 
 class MeshBlock {
     public:
@@ -42,7 +43,7 @@ __host__ __device__ void MeshBlock::ImportNumpyData(const std::string npy_path) 
     npy::npy_data d = npy::read_npy<float>(npy_path);
     std::vector<float> npy_data = d.data;
     float *p_data = npy_data.data();
-    std::vector<unsigned long> shape = d.shape;
+    std::vector<int> shape = d.shape;
     size_t bytes = npy_data.size() * sizeof(float);
 
     printf("finished numpy import");
